+++ packages/build-clean/src/errors/index.ts
@@ -1,0 +1,10 @@
+import { BuildCleanError } from './BuildCleanError';
+import { CleanRefusedError } from './CleanRefusedError';
+import { MissingMetafileError } from './MissingMetafileError';
+import { MissingOutputDirectoryError } from './MissingOutputDirectoryError';
+import { OutputDirectoryContainsBaseError } from './OutputDirectoryContainsBaseError';
+import { OutputDirectoryIsBaseError } from './OutputDirectoryIsBaseError';
+import { OutputDirectoryIsSourceError } from './OutputDirectoryIsSourceError';
+import { OutputDirectoryOutsideBaseError } from './OutputDirectoryOutsideBaseError';
+
+export { BuildCleanError, CleanRefusedError, MissingMetafileError, MissingOutputDirectoryError, OutputDirectoryContainsBaseError, OutputDirectoryIsBaseError, OutputDirectoryIsSourceError, OutputDirectoryOutsideBaseError };