+++ packages/build-clean/src/index.ts
@@ -1,1 +1,4 @@
-export { plugin as default } from './core/plugin';
+import { plugin } from './core/plugin';
+import { BuildCleanError, CleanRefusedError, MissingMetafileError, MissingOutputDirectoryError, OutputDirectoryContainsBaseError, OutputDirectoryIsBaseError, OutputDirectoryIsSourceError, OutputDirectoryOutsideBaseError } from './errors';
+
+export { BuildCleanError, CleanRefusedError, MissingMetafileError, MissingOutputDirectoryError, OutputDirectoryContainsBaseError, OutputDirectoryIsBaseError, OutputDirectoryIsSourceError, OutputDirectoryOutsideBaseError, plugin as default };